# shortlist/server/api/requests.py
"""The Sonarr/Radarr approval inbox: list wanted-but-missing titles, send the chosen ones, reject the rest.

A request asks a download app for a file — it touches no Plex object. It is gated only on the owner
session and on requests being configured. Sending runs in a
worker thread (the Arr/TMDB clients are sync) and respects ``dry_run``.
"""

from __future__ import annotations

import asyncio
from datetime import UTC, datetime

import httpx
from fastapi import APIRouter, Depends, HTTPException, Request
from loguru import logger
from pydantic import BaseModel, Field

from shortlist.engine.models import MediaType, MissingTitle
from shortlist.engine.requests import request_titles
from shortlist.server.auth import require_owner
from shortlist.server.db.models import Event, RequestCandidate, iso_utc

router = APIRouter(prefix="/requests", tags=["requests"], dependencies=[Depends(require_owner)])

# Pending first (the owner's to-do list), then sent, then rejected — so the inbox opens on what needs a decision.
_STATUS_ORDER = {"pending": 0, "sent": 1, "rejected": 2}


class RequestWhyOut(BaseModel):
    user: str  # whose taste surfaced it
    row: str  # the row that wanted it (the name the user sees)
    seed: str  # the history title behind it ("because you watched …"); "" for seedless sources
    source: str  # the candidate source that produced it


class RequestCandidateOut(BaseModel):
    id: int
    tmdb_id: int
    media_type: str
    title: str
    year: int | None
    imdb_id: str = ""  # "tt…" for a direct IMDb link; "" -> the UI falls back to an IMDb search
    # TMDB poster path ("/abc.jpg"). The UI builds the image URL and its size; "" -> placeholder tile.
    poster_path: str = ""
    rating: float
    vote_count: int
    demand: int
    tags: list[str]
    wanters: list[str]
    why: list[RequestWhyOut]  # per (person, row) provenance — which row, and why it got here
    status: str
    detail: str
    excluded: bool = False  # on a Sonarr/Radarr exclusion list — the inbox warns approving is a no-op
    arr_slug: str | None = None  # the arr titleSlug -> the sent log deep-links straight to its page
    updated_at: str | None  # when this row last changed state (the "sent at" for a sent item)
    # Live Arr download status is fetched separately via GET /requests/status (one round-trip for the
    # whole inbox) and merged in the UI — it is NOT carried on the list payload, which would force an
    # Arr call per row on every list fetch.


class RequestAction(BaseModel):
    #: Bounded because every handler feeds this straight into `.in_()`. SQLite's compiled parameter
    #: ceiling (SQLITE_MAX_VARIABLE_NUMBER, 999 on older builds) turns an over-long list into an
    #: OperationalError — a 500 with a SQL string in it — rather than a refusal the caller can read.
    ids: list[int] = Field(max_length=1000)
    dry_run: bool = False


#: Hard ceiling on one inbox read. The sent log only grows — every run that wants a title the library
#: lacks adds a row — so an unbounded read is a query that gets slower for ever and eventually times
#: out the page. Pending is what the owner acts on and is self-limiting (you clear it); the tail is
#: history, and the sort puts pending first, so a cap can only ever truncate the oldest history.
MAX_INBOX = 500


@router.get("")
def list_requests(request: Request) -> list[RequestCandidateOut]:
    """The whole inbox: pending first (most-wanted, best-rated on top), then sent, then rejected.

    Rows the owner cleared from the Sent log (``hidden``) are excluded — they stay in the DB as sent
    tombstones (so the title isn't re-requested) but never show in the UI again.

    Capped at :data:`MAX_INBOX`. The cap is applied AFTER the status sort, in Python, because the
    ordering is by (status, demand, rating) and a SQL LIMIT before that sort would cut arbitrary rows
    rather than the oldest history.
    """
    with request.app.state.sessions() as session:
        rows = session.query(RequestCandidate).filter(~RequestCandidate.hidden).all()
    rows.sort(key=lambda r: (_STATUS_ORDER.get(r.status, 9), -r.demand, -r.rating))
    rows = rows[:MAX_INBOX]
    return [
        RequestCandidateOut(
            id=r.id,
            tmdb_id=r.tmdb_id,
            media_type=r.media_type,
            title=r.title,
            year=r.year,
            imdb_id=r.imdb_id or "",
            poster_path=r.poster_path or "",
            rating=r.rating,
            vote_count=r.vote_count,
            demand=r.demand,
            tags=list(r.tags or []),
            wanters=list(r.wanters or []),
            why=[RequestWhyOut(**w) for w in (r.why or [])],
            status=r.status,
            detail=r.detail,
            excluded=bool(r.excluded),
            arr_slug=r.arr_slug,
            updated_at=iso_utc(r.updated_at),
        )
        for r in rows
    ]


@router.post("/reject")
def reject_requests(body: RequestAction, request: Request) -> dict:
    """Permanently dismiss the given titles.

    A rejected title is kept on file as a tombstone: it leaves the pending list AND every later run
    skips re-queuing it (``_persist_request_queue`` only touches ``pending`` rows), so a dismissed
    suggestion can never come back on its own. Use ``/delete`` instead to remove a title without
    blocking it — or to lift a rejection so a future run may surface it again.
    """
    with request.app.state.sessions() as session:
        rows = session.query(RequestCandidate).filter(RequestCandidate.id.in_(body.ids)).all()
        for row in rows:
            row.status = "rejected"
        session.add(Event(scope="requests.reject", level="info", message={"ids": body.ids, "count": len(rows)}))
        session.commit()
    return {"rejected": len(rows)}


@router.post("/restore")
def restore_requests(body: RequestAction, request: Request) -> dict:
    """Un-reject: move rejected titles back to the pending queue (Waiting) so they can be sent again.

    Only ``rejected`` rows are restored; ``pending``/``sent`` are left as they are. The row keeps its
    recorded demand/wanters/why/tags, so it reappears in Waiting exactly as it was, ready to send —
    unlike a run, which would only re-surface it if the same taste turned it up again.
    """
    with request.app.state.sessions() as session:
        rows = (
            session.query(RequestCandidate)
            .filter(RequestCandidate.id.in_(body.ids), RequestCandidate.status == "rejected")
            .all()
        )
        for row in rows:
            row.status = "pending"
        session.add(Event(scope="requests.restore", level="info", message={"ids": body.ids, "count": len(rows)}))
        session.commit()
    return {"restored": len(rows)}


@router.post("/delete")
def delete_requests(body: RequestAction, request: Request) -> dict:
    """Remove the given titles from the inbox entirely, leaving no trace.

    Unlike ``/reject`` (a permanent tombstone), a deleted row is gone — so if a later run's picks turn
    up the same title again, it returns to the pending queue. Two uses: clear a title off the list
    without blocking it forever, or delete a previously *rejected* title to let it come back.

    ``sent`` rows are never deleted: that status is a load-bearing tombstone (``_persist_request_queue``)
    that stops a still-downloading title from being seen as "missing" and re-requested every night.
    Dropping it would resurrect that bug, so a ``sent`` id in the request is skipped, not deleted.
    """
    with request.app.state.sessions() as session:
        rows = (
            session.query(RequestCandidate)
            .filter(RequestCandidate.id.in_(body.ids), RequestCandidate.status != "sent")
            .all()
        )
        count = len(rows)
        for row in rows:
            session.delete(row)
        session.add(Event(scope="requests.delete", level="info", message={"ids": body.ids, "count": count}))
        session.commit()
    return {"deleted": count}


@router.post("/clear")
def clear_requests(body: RequestAction, request: Request) -> dict:
    """Clear the given SENT titles from the send log — hide them, don't delete them.

    A sent row is a load-bearing tombstone: dropping it lets a still-downloading title look "missing"
    and get re-requested every night (see ``delete_requests``). So "clear" sets ``hidden`` instead —
    the row stays ``sent`` and keeps protecting against re-request, but never shows in the inbox again.
    Only ``sent`` rows are cleared; a pending/rejected id is ignored (those have Delete / Reject).
    """
    with request.app.state.sessions() as session:
        rows = (
            session.query(RequestCandidate)
            .filter(RequestCandidate.id.in_(body.ids), RequestCandidate.status == "sent")
            .all()
        )
        count = 0
        for row in rows:
            if not row.hidden:
                row.hidden = True
                count += 1
        session.add(Event(scope="requests.clear", level="info", message={"ids": body.ids, "count": count}))
        session.commit()
    return {"cleared": count}


@router.get("/status")
async def get_arr_status(request: Request) -> dict[int, str | None]:
    """Arr download status for every request row. Returns {request_id: 'downloaded' | 'downloading' | ...}.

    Covers waiting rows as well as sent ones. A waiting title is normally absent from the Arrs — the
    nightly pass drops anything they already track — so a status there means the owner (or another
    tool) added it by hand since, which is exactly the case where "why is this still waiting?" needs
    an answer. Rejected rows are skipped: nothing is going to happen to them.

    Whole-library maps, not per-title lookups, so the cost is a handful of calls no matter how long
    the inbox is. Runs in an executor since the Arr clients are sync. A title neither app tracks
    appears as None.
    """
    state = request.app.state
    svc = state.run_service

    def _fetch_statuses() -> dict[int, str | None]:
        cfg, tmdb = svc.build_requests_context()
        if cfg is None:
            return {}

        from shortlist.engine.clients.arr import RadarrClient, SonarrClient

        with state.sessions() as session:
            rows = session.query(RequestCandidate).filter(RequestCandidate.status.in_(("pending", "sent"))).all()

        # One fetch per app up front. A failure here is not fatal: the inbox simply shows no status
        # rather than erroring, which is what it did before this endpoint existed.
        movies: dict[int, str] = {}
        shows_by_tvdb: dict[int, str] = {}
        shows_by_tmdb: dict[int, str] = {}
        if cfg.radarr:
            try:
                movies = RadarrClient(cfg.radarr).status_by_tmdb()
            except Exception as e:
                logger.warning("request status: Radarr lookup failed ({})", e)
        if cfg.sonarr:
            try:
                shows_by_tvdb, shows_by_tmdb = SonarrClient(cfg.sonarr).status_by_ids()
            except Exception as e:
                logger.warning("request status: Sonarr lookup failed ({})", e)

        statuses: dict[int, str | None] = {}
        for row in rows:
            if row.media_type == "movie":
                statuses[row.id] = movies.get(row.tmdb_id)
                continue
            # Sonarr v4 carries tmdbId on every series, so the map answers directly. On v3 it doesn't,
            # and only then is a per-title TMDB→TVDB lookup worth paying for (cached in the client).
            status = shows_by_tmdb.get(row.tmdb_id)
            if status is None and shows_by_tvdb and not shows_by_tmdb:
                try:
                    tvdb_id = tmdb.external_ids(row.tmdb_id, MediaType.SHOW).get("tvdb_id")
                # Deliberately NOT a bare `except Exception`: this used to pass `MediaType.TV`, which
                # does not exist, and the AttributeError was swallowed to a debug line — so on Sonarr
                # v3 the fallback silently no-op'd for ever and every show showed a blank status. Only
                # a transport failure or the TMDB client's own HTTP error is tolerable here; anything
                # else is a bug and must be loud.
                except (httpx.HTTPError, RuntimeError) as e:
                    logger.debug("request status: tvdb lookup for {!r} failed ({})", row.title, e)
                    tvdb_id = None
                status = shows_by_tvdb.get(tvdb_id) if tvdb_id else None
            statuses[row.id] = status

        return statuses

    return await asyncio.get_running_loop().run_in_executor(None, _fetch_statuses)


@router.post("/send")
async def send_requests(body: RequestAction, request: Request) -> dict:
    """Ask Sonarr/Radarr for the chosen pending titles.

    A dry run previews the outcomes without asking and leaves every row pending. A real send marks a
    row ``sent`` only when the app accepted it; a skip/error leaves it pending with the reason recorded,
    so the owner can see why it didn't go and try again.
    """
    state = request.app.state
    svc = state.run_service

    def _send() -> dict:
        cfg, tmdb = svc.build_requests_context()
        if cfg is None:
            raise HTTPException(status_code=409, detail="Turn on Sonarr/Radarr requests in Settings first.")
        with state.sessions() as session:
            rows = (
                session.query(RequestCandidate)
                .filter(RequestCandidate.id.in_(body.ids), RequestCandidate.status == "pending")
                .all()
            )
            titles = [
                MissingTitle(
                    tmdb_id=row.tmdb_id,
                    title=row.title,
                    media_type=MediaType(row.media_type),
                    year=row.year,
                    rating=row.rating,
                    vote_count=row.vote_count,
                    demand=row.demand,
                    tags=set(row.tags or []),
                )
                for row in rows
            ]
            report = request_titles(cfg, tmdb, titles, dry_run=body.dry_run)
            by_key = {(o.tmdb_id, o.media_type.value): o for o in report.outcomes}
            outcomes = []
            for row in rows:
                outcome = by_key.get((row.tmdb_id, row.media_type))
                if outcome is None:
                    continue
                row.detail = outcome.detail
                if outcome.arr_slug:
                    row.arr_slug = outcome.arr_slug  # so the sent log deep-links to the arr page
                if not body.dry_run and outcome.status == "requested":
                    row.status = "sent"
                    # Stamped once, here, so "watched since sent" can compare against the real send
                    # time rather than an `updated_at` that later edits move around.
                    row.sent_at = datetime.now(UTC)
                outcomes.append({"id": row.id, "title": row.title, "status": outcome.status, "detail": outcome.detail})
            session.add(
                Event(scope="requests.send", level="info", message={"dry_run": body.dry_run, "outcomes": outcomes})
            )
            session.commit()
            sent = sum(1 for o in outcomes if o["status"] in ("requested", "would_request"))
            return {"sent": sent, "dry_run": body.dry_run, "outcomes": outcomes}

    return await asyncio.get_running_loop().run_in_executor(None, _send)