# web/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestAction": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 1000,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "RequestAction",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shortlist",
    "version": "0.1.0b8"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/requests/clear": {
      "post": {
        "description": "Clear the given SENT titles from the send log \u2014 hide them, don't delete them.\n\nA sent row is a load-bearing tombstone: dropping it lets a still-downloading title look \"missing\"\nand get re-requested every night (see ``delete_requests``). So \"clear\" sets ``hidden`` instead \u2014\nthe row stays ``sent`` and keeps protecting against re-request, but never shows in the inbox again.\nOnly ``sent`` rows are cleared; a pending/rejected id is ignored (those have Delete / Reject).",
        "operationId": "clear_requests_api_requests_clear_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clear Requests Api Requests Clear Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Requests",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/requests/delete": {
      "post": {
        "description": "Remove the given titles from the inbox entirely, leaving no trace.\n\nUnlike ``/reject`` (a permanent tombstone), a deleted row is gone \u2014 so if a later run's picks turn\nup the same title again, it returns to the pending queue. Two uses: clear a title off the list\nwithout blocking it forever, or delete a previously *rejected* title to let it come back.\n\n``sent`` rows are never deleted: that status is a load-bearing tombstone (``_persist_request_queue``)\nthat stops a still-downloading title from being seen as \"missing\" and re-requested every night.\nDropping it would resurrect that bug, so a ``sent`` id in the request is skipped, not deleted.",
        "operationId": "delete_requests_api_requests_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Requests Api Requests Delete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Requests",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/requests/reject": {
      "post": {
        "description": "Permanently dismiss the given titles.\n\nA rejected title is kept on file as a tombstone: it leaves the pending list AND every later run\nskips re-queuing it (``_persist_request_queue`` only touches ``pending`` rows), so a dismissed\nsuggestion can never come back on its own. Use ``/delete`` instead to remove a title without\nblocking it \u2014 or to lift a rejection so a future run may surface it again.",
        "operationId": "reject_requests_api_requests_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Requests Api Requests Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Requests",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/requests/restore": {
      "post": {
        "description": "Un-reject: move rejected titles back to the pending queue (Waiting) so they can be sent again.\n\nOnly ``rejected`` rows are restored; ``pending``/``sent`` are left as they are. The row keeps its\nrecorded demand/wanters/why/tags, so it reappears in Waiting exactly as it was, ready to send \u2014\nunlike a run, which would only re-surface it if the same taste turned it up again.",
        "operationId": "restore_requests_api_requests_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Restore Requests Api Requests Restore Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Requests",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/requests/send": {
      "post": {
        "description": "Ask Sonarr/Radarr for the chosen pending titles.\n\nA dry run previews the outcomes without asking and leaves every row pending. A real send marks a\nrow ``sent`` only when the app accepted it; a skip/error leaves it pending with the reason recorded,\nso the owner can see why it didn't go and try again.",
        "operationId": "send_requests_api_requests_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Send Requests Api Requests Send Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Requests",
        "tags": [
          "requests"
        ]
      }
    },
...
  }
}